# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/api-keys": {
      "post": {
        "operationId": "AuthController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreatedResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new API key (admin only)",
        "tags": [
          "auth"
        ]
      },
      "get": {
        "operationId": "AuthController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All API keys (the plaintext key is never returned; only the keyPrefix).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all API keys (admin only)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/api-keys/{id}": {
      "get": {
        "operationId": "AuthController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The API key (plaintext never returned; only the keyPrefix).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get API key details (admin only)",
        "tags": [
          "auth"
        ]
      },
      "put": {
        "operationId": "AuthController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The change would remove the last usable admin key."
          }
        },
        "summary": "Update API key (admin only)",
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "API key deleted"
          },
          "409": {
            "description": "The key is the last usable admin key."
          }
        },
        "summary": "Delete API key (admin only)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/api-keys/{id}/revoke": {
      "post": {
        "operationId": "AuthController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The revoked API key (isActive now false).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The key is the last usable admin key."
          }
        },
        "summary": "Revoke API key (admin only)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/validate": {
      "post": {
        "operationId": "AuthValidateController_validate",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API key to validate",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key is valid"
          },
          "401": {
            "description": "Invalid or missing API key"
          }
        },
        "summary": "Validate an API key",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.14.2",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "This instance (the origin serving these docs)"
    },
    {
      "url": "http://{host}:{port}",
      "description": "Another instance (set host and port)",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
      "CreateApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Friendly name for the API key",
            "example": "Production Bot"
          },
          "role": {
            "type": "string",
            "description": "Role/permission level",
            "enum": [
              "admin",
              "operator",
              "viewer"
            ],
            "default": "operator"
          },
          "allowedIps": {
            "description": "Allowed IP addresses (whitelist)",
            "example": [
              "192.168.1.1",
              "10.0.0.0/8"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedSessions": {
            "description": "Allowed session IDs this key can access",
            "example": [
              "session-uuid-1",
              "session-uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expiresAt": {
            "type": "string",
            "description": "Expiration date (ISO 8601)",
            "example": "2027-12-31T23:59:59Z"
          }
        },
        "required": [
          "name"
        ]
      },
      "ApiKeyCreatedResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string",
            "description": "First 8 characters of the key (for identification)"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "operator",
              "viewer"
            ]
          },
          "allowedIps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedSessions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastUsedAt": {
            "format": "date-time",
            "type": "string"
          },
          "usageCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "apiKey": {
            "type": "string",
            "description": "Full API key (only shown once at creation)",
            "example": "owa_k1_abc123..."
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "role",
          "isActive",
          "usageCount",
          "createdAt",
          "apiKey"
        ]
      },
      "ApiKeyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string",
            "description": "First 8 characters of the key (for identification)"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "operator",
              "viewer"
            ]
          },
          "allowedIps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedSessions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastUsedAt": {
            "format": "date-time",
            "type": "string"
          },
          "usageCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "role",
          "isActive",
          "usageCount",
          "createdAt"
        ]
      },
      "UpdateApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "operator",
              "viewer"
            ]
          },
          "allowedIps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowedSessions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expiresAt": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}